+++ packages/htn/__tests__/mocks/index.ts
@@ -17,5 +17,4 @@ export { FryTask } from './primitives/fry-task';
 export { MakeTomatoSauceTask } from './primitives/make-tomato-sauce-task';
 export { ServeDishTask } from './primitives/serve-dish-task';
 
-export { MockedAgent, MockedAgentFactory } from './implementations';
 export { type KitchenState, createKitchen } from './state';
